Add a software AX mixer and compile the SDK's AX voice/parameter code
src/port/pc_ax.c stands in for AX.c, AXOut, AXCL, AXAux and the DSP: 64
voices mixed in 5 ms frames at 32 kHz (DSP-ADPCM / PCM16 / PCM8, 16.16 SRC
with linear interpolation, 1.15 volume envelopes, main plus two aux buses),
submitted to SDL from the frame end and from the load-wait ticks. The
accelerator wraps only on equality with the end address, which HAL's
streamer relies on when a block rotates into a higher ARAM slot.

The SDK's AXAlloc/AXVPB/AXSPB/AXProf and the AXFX delay compile as-is once
their `long`s are `s32` (identical on the console, 8 bytes here: the delay
line zero-fill overran its heap). AXVPB.c and AXAlloc.c are included via
src/port/ax_*_glue.c for two PC-only hooks: PC_AX_SYNC_FIX copies the
whole addr block before the partial setters (synchronous file callbacks
put AXSetVoiceAddr and AXSetVoiceLoopAddr in the same frame, and the SDK's
else-if dropped loopFlag/format), and PC_AX_HOOKS traces frees. AXFX
reverb/chorus are MWCC inline asm and stay stubbed.

MELEE_AXTRACE=1 traces voices/streams/scripts, MELEE_NO_AUDIO=1 mutes,
MELEE_AUDIO_DUMP=<path> writes raw 32 kHz s16 stereo.

+++ extern/dolphin/src/dolphin/ax/AXSPB.c
@@ -8,25 +8,25 @@
 static struct _AXSPB __AXStudio ATTRIBUTE_ALIGN(32);
 
 // .sbss
-static long __AXSpbAL;
-static long __AXSpbAR;
-static long __AXSpbAS;
-static long __AXSpbAAL;
-static long __AXSpbAAR;
-static long __AXSpbAAS;
-static long __AXSpbABL;
-static long __AXSpbABR;
-static long __AXSpbABS;
+static s32 __AXSpbAL;
+static s32 __AXSpbAR;
+static s32 __AXSpbAS;
+static s32 __AXSpbAAL;
+static s32 __AXSpbAAR;
+static s32 __AXSpbAAS;
+static s32 __AXSpbABL;
+static s32 __AXSpbABR;
+static s32 __AXSpbABS;
 
 u32 __AXGetStudio(void)
 {
     return (u32) &__AXStudio;
 }
 
-void __AXDepopFade(long* hostSum, long* dspVolume, s16* dspDelta)
+void __AXDepopFade(s32* hostSum, s32* dspVolume, s16* dspDelta)
 {
     int frames;
-    long delta;
+    s32 delta;
 
     frames = *hostSum / 160;
