+++ extern/dolphin/src/dolphin/axfx/delay.c
@@ -5,16 +5,16 @@
 void AXFXDelayCallback(struct AXFX_BUFFERUPDATE* bufferUpdate,
                        struct AXFX_DELAY* delay)
 {
-    long l;
-    long r;
-    long s;
-    long* lBuf;
-    long* rBuf;
-    long* sBuf;
+    s32 l;
+    s32 r;
+    s32 s;
+    s32* lBuf;
+    s32* rBuf;
+    s32* sBuf;
     u32 i;
-    long* left;
-    long* right;
-    long* sur;
+    s32* left;
+    s32* right;
+    s32* sur;
 
     left = bufferUpdate->left;
     right = bufferUpdate->right;
@@ -44,10 +44,10 @@ void AXFXDelayCallback(struct AXFX_BUFFERUPDATE* bufferUpdate,
 
 int AXFXDelaySettings(struct AXFX_DELAY* delay)
 {
-    unsigned long i;
-    long* l;
-    long* r;
-    long* s;
+    u32 i;
+    s32* l;
+    s32* r;
+    s32* s;
     int old;
 
     AXFXDelayShutdown(delay);
